+++ docs/pages/Documentation.vue
@@ -2,8 +2,8 @@
     <article class="content">
         <p>
             Buefy (pronounced /ˈbyo͞oˌfī/, like <b>beau-fy</b>) is a user interface component library made on top
-            of two <em>fresh</em> but very well known frameworks, them being
-            <a href="https://vuejs.org" target="_blank">Vue.js</a> and <a href="https://bulma.io/">Bulma</a>.
+            of <a href="https://vuejs.org" target="_blank">Vue.js</a> and <a href="https://bulma.io/">Bulma</a>;
+            two <em>fresh</em>, but well-known, frameworks.
             Buefy has two core principles: 1. keep stuff simple and 2. be lightweight.
         </p>