+++ docs/pages/components/message/examples/ExAutoCloseWithProgressBar.vue
@@ -1,10 +1,16 @@
 <template>
     <section>
         <b-button
-            label="Toggle"
-            @click="isActive = !isActive" />
+            label="Show"
+            :disabled="isActive"
+            @click="isActive = true" />
         <b-message 
-            title="Default" 
+            auto-close 
+            :progress-bar='progressBar' 
+            :duration="duration" 
+            title="Error!" 
+            type="is-dark" 
+            has-icon 
             v-model="isActive" 
             aria-close-label="Close message">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
@@ -16,7 +22,9 @@
     export default {
         data() {
             return {
-                isActive: true
+                isActive: false,
+                duration: 4000,
+                progressBar: true
             }
         }
     }