+++ docs/pages/components/notification/examples/ExSimple.vue
@@ -4,7 +4,9 @@
             label="Toggle"
             class="block"
             @click="isActive = !isActive" />
-        <b-notification v-model="isActive" aria-close-label="Close notification">
+        <b-notification 
+            v-model="isActive" 
+            aria-close-label="Close notification">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
         </b-notification>
     </section>