+++ docs/pages/components/upload/examples/ExRounded.vue
@@ -4,7 +4,7 @@
       <b-field class="file is-primary" :class="{'has-name': !!file}">
         <b-upload v-model="file" class="file-label" rounded>
           <span class="file-cta">
-            <b-icon icon="upload"></b-icon>
+            <b-icon class="file-icon" icon="upload"></b-icon>
             <span class="file-label">{{ file.name || "Click to upload"}}</span>
           </span>
         </b-upload>