+++ src/components/autocomplete/Autocomplete.vue
@@ -3,7 +3,7 @@
         <b-input
             v-model="newValue"
             ref="input"
-            type="text"
+            :type="type"
             :size="size"
             :loading="loading"
             :rounded="rounded"
@@ -14,11 +14,11 @@
             :maxlength="maxlength"
             :autocomplete="newAutocomplete"
             :use-html5-validation="false"
+            :aria-autocomplete="ariaAutocomplete"
             v-bind="$attrs"
             @input="onInput"
             @focus="focused"
             @blur="onBlur"
-            @keyup.native.esc.prevent="isActive = false"
             @keydown.native="keydown"
             @keydown.native.up.prevent="keyArrows('up')"
             @keydown.native.down.prevent="keyArrows('down')"
@@ -38,7 +38,14 @@
                     class="dropdown-content"
                     v-show="isActive"
                     :style="contentStyle">
-                    <div v-if="hasHeaderSlot" class="dropdown-item">
+                    <div
+                        v-if="hasHeaderSlot"
+                        class="dropdown-item dropdown-header"
+                        role="button"
+                        tabindex="0"
+                        :class="{ 'is-hovered': headerHovered }"
+                        @click="(event) => checkIfHeaderOrFooterSelected(event, true)"
+                    >
                         <slot name="header" />
                     </div>
                     <template v-for="(element, groupindex) in computedData">
@@ -78,7 +85,14 @@
                         class="dropdown-item is-disabled">
                         <slot name="empty" />
                     </div>
-                    <div v-if="hasFooterSlot" class="dropdown-item">
+                    <div
+                        v-if="hasFooterSlot"
+                        class="dropdown-item dropdown-footer"
+                        role="button"
+                        tabindex="0"
+                        :class="{ 'is-hovered': footerHovered }"
+                        @click="(event) => checkIfHeaderOrFooterSelected(event, true)"
+                    >
                         <slot name="footer" />
                     </div>
                 </div>
@@ -121,6 +135,7 @@ export default {
         customFormatter: Function,
         checkInfiniteScroll: Boolean,
         keepOpen: Boolean,
+        selectOnClickOutside: Boolean,
         clearable: Boolean,
         maxHeight: [String, Number],
         dropdownPosition: {
@@ -132,18 +147,27 @@ export default {
         iconRight: String,
         iconRightClickable: Boolean,
         appendToBody: Boolean,
+        type: {
+            type: String,
+            default: 'text'
+        },
         confirmKeys: {
             type: Array,
             default: () => ['Tab', 'Enter']
-        }
+        },
+        selectableHeader: Boolean,
+        selectableFooter: Boolean
     },
     data() {
         return {
             selected: null,
             hovered: null,
+            headerHovered: null,
+            footerHovered: null,
             isActive: false,
             newValue: this.value,
             newAutocomplete: this.autocomplete || 'off',
+            ariaAutocomplete: this.keepFirst ? 'both' : 'list',
             isListInViewportVertically: true,
             hasFocus: false,
             style: {},
@@ -331,6 +355,8 @@ export default {
                 this.$nextTick(() => {
                     if (this.isActive) {
                         this.selectFirstOption(this.computedData)
+                    } else {
+                        this.setHovered(null)
                     }
                 })
             }
@@ -352,11 +378,15 @@ export default {
          */
         setSelected(option, closeDropdown = true, event = undefined) {
             if (option === undefined) return
-
             this.selected = option
             this.$emit('select', this.selected, event)
             if (this.selected !== null) {
-                this.newValue = this.clearOnSelect ? '' : this.getValue(this.selected)
+                if (this.clearOnSelect) {
+                    const input = this.$refs.input.$refs.input
+                    input.value = ''
+                } else {
+                    this.newValue = this.getValue(this.selected)
+                }
                 this.setHovered(null)
             }
             closeDropdown && this.$nextTick(() => {
@@ -368,14 +398,14 @@ export default {
         /**
          * Select first option
          */
-        selectFirstOption(element) {
+        selectFirstOption(computedData) {
             this.$nextTick(() => {
-                if (element.length) {
-                    // If has visible data or open on focus, keep updating the hovered
-                    const option = element[0].items[0]
-                    if (this.openOnFocus || (this.newValue !== '' && this.hovered !== option)) {
-                        this.setHovered(option)
-                    }
+                const nonEmptyElements = computedData.filter(
+                    (element) => element.items && element.items.length
+                )
+                if (nonEmptyElements.length) {
+                    const option = nonEmptyElements[0].items[0]
+                    this.setHovered(option)
                 } else {
                     this.setHovered(null)
                 }
@@ -387,15 +417,40 @@ export default {
             // prevent emit submit event
             if (key === 'Enter') event.preventDefault()
             // Close dropdown on Tab & no hovered
-            this.isActive = key !== 'Tab'
-            if (this.hovered === null) return
+            if (key === 'Escape' || key === 'Tab') {
+                this.isActive = false
+            }
+
             if (this.confirmKeys.indexOf(key) >= 0) {
                 // If adding by comma, don't add the comma to the input
                 if (key === ',') event.preventDefault()
-
                 // Close dropdown on select by Tab
                 const closeDropdown = !this.keepOpen || key === 'Tab'
+                if (this.hovered === null) {
+                    // header and footer uses headerHovered && footerHovered. If header or footer
+                    // was selected then fire event otherwise just return so a value isn't selected
+                    this.checkIfHeaderOrFooterSelected(event, false, closeDropdown)
+                    return
+                }
                 this.setSelected(this.hovered, closeDropdown, event)
+            }
+        },
+
+        /**
+         * Check if header or footer was selected.
+         */
+        checkIfHeaderOrFooterSelected(event, triggeredByclick, closeDropdown = true) {
+            if (this.selectableHeader && (this.headerHovered || triggeredByclick)) {
+                this.$emit('select-header', event)
+                this.headerHovered = false
+                if (triggeredByclick) this.setHovered(null)
+                if (closeDropdown) this.isActive = false
+            }
+            if (this.selectableFooter && (this.footerHovered || triggeredByclick)) {
+                this.$emit('select-footer', event)
+                this.footerHovered = false
+                if (triggeredByclick) this.setHovered(null)
+                if (closeDropdown) this.isActive = false
             }
         },
 
@@ -405,7 +460,7 @@ export default {
         clickedOutside(event) {
             const target = isCustomElement(this) ? event.composedPath()[0] : event.target
             if (!this.hasFocus && this.whiteList.indexOf(target) < 0) {
-                if (this.keepFirst && this.hovered) {
+                if (this.keepFirst && this.hovered && this.selectOnClickOutside) {
                     this.setSelected(this.hovered, true)
                 } else {
                     this.isActive = false
@@ -469,14 +524,44 @@ export default {
             if (this.isActive) {
                 const data = this.computedData.map(
                     (d) => d.items).reduce((a, b) => ([...a, ...b]), [])
-                let index = data.indexOf(this.hovered) + sum
+                if (this.hasHeaderSlot && this.selectableHeader) {
+                    data.unshift(undefined)
+                }
+                if (this.hasFooterSlot && this.selectableFooter) {
+                    data.push(undefined)
+                }
+
+                let index
+                if (this.headerHovered) {
+                    index = 0 + sum
+                } else if (this.footerHovered) {
+                    index = (data.length - 1) + sum
+                } else {
+                    index = data.indexOf(this.hovered) + sum
+                }
+
                 index = index > data.length - 1 ? data.length - 1 : index
                 index = index < 0 ? 0 : index
 
-                this.setHovered(data[index])
+                this.footerHovered = false
+                this.headerHovered = false
+                this.setHovered(data[index] !== undefined ? data[index] : null)
+                if (this.hasFooterSlot && this.selectableFooter && index === data.length - 1) {
+                    this.footerHovered = true
+                }
+                if (this.hasHeaderSlot && this.selectableHeader && index === 0) {
+                    this.headerHovered = true
+                }
 
                 const list = this.$refs.dropdown.querySelector('.dropdown-content')
-                const element = list.querySelectorAll('a.dropdown-item:not(.is-disabled)')[index]
+                let querySelectorText = 'a.dropdown-item:not(.is-disabled)'
+                if (this.hasHeaderSlot && this.selectableHeader) {
+                    querySelectorText += ',div.dropdown-header'
+                }
+                if (this.hasFooterSlot && this.selectableFooter) {
+                    querySelectorText += ',div.dropdown-footer'
+                }
+                const element = list.querySelectorAll(querySelectorText)[index]
 
                 if (!element) return
 
@@ -504,6 +589,7 @@ export default {
             if (this.openOnFocus) {
                 this.isActive = true
                 if (this.keepFirst) {
+                    // If open on focus, update the hovered
                     this.selectFirstOption(this.computedData)
                 }
             }
@@ -518,7 +604,7 @@ export default {
             this.hasFocus = false
             this.$emit('blur', event)
         },
-        onInput(event) {
+        onInput() {
             const currentValue = this.getValue(this.selected)
             if (currentValue && currentValue === this.newValue) return
             this.$emit('typing', this.newValue)