+++ src/components/notification/NotificationNotice.vue
@@ -1,7 +1,11 @@
 <template>
     <b-notification
         v-bind="$options.propsData"
-        @close="close">
+        ref="notification"
+        @click="click"
+        @close="close"
+        @mouseenter.native="pause"
+        @mouseleave.native="removePause">
         <slot />
     </b-notification>
 </template>
@@ -9,6 +13,7 @@
 <script>
 import config from '../../utils/config'
 import NoticeMixin from '../../utils/NoticeMixin.js'
+import { removeElement } from '../../utils/helpers'
 
 export default {
     name: 'BNotificationNotice',
@@ -16,6 +21,21 @@ export default {
     data() {
         return {
             newDuration: this.duration || config.defaultNotificationDuration
+        }
+    },
+    methods: {
+        close() {
+            if (!this.isPaused) {
+                clearTimeout(this.timer)
+                this.$refs.notification.isActive = false
+                this.$emit('close')
+
+                // Timeout for the animation complete before destroying
+                setTimeout(() => {
+                    this.$destroy()
+                    removeElement(this.$el)
+                }, 150)
+            }
         }
     }
 }