+++ src/components/taginput/Taginput.vue
@@ -291,7 +291,11 @@ export default {
                 }
             }
 
-            this.newTag = ''
+            // after autocomplete events
+            requestAnimationFrame(() => {
+                this.newTag = ''
+                this.$emit('typing', '')
+            })
         },
 
         getNormalizedTagText(tag) {
@@ -313,9 +317,6 @@ export default {
             if (!option) return
 
             this.addTag(option)
-            this.$nextTick(() => {
-                this.newTag = ''
-            })
         },
 
         removeTag(index, event) {